Former-commit-id: f1fc7abadf30ca94c75bac468f8da2f5b7cfbdc0

+++ Sources/Internal/Render/OcclusionQuery.h
@@ -95,7 +95,7 @@ private:
     {
         OcclusionQuery query;
         uint32 next;
-        uint32 salt;
+        uint16 salt;
     };
     Vector<OcclusionQueryItem> queries;
 };
